Fix folding to include the final layer. Also fixed import errors from renaming
