+++ src/vit_prisma/models/layers/mlp.py
@@ -1,5 +1,5 @@
 import torch.nn as nn
-from vit_prisma.prisma.hook_point import HookPoint
+from vit_prisma.prisma_tools.hook_point import HookPoint
 from vit_prisma.configs.HookedViTConfig import HookedViTConfig
 
 from typing import Dict, Optional, Tuple, Union
