+++ src/vit_prisma/prisma_tools/hooked_root_module.py
@@ -4,7 +4,7 @@ from typing import Dict, Optional, Tuple, Union, Callable, Sequence, List
 from jaxtyping import Float, Int
 
 import torch.nn as nn 
-from vit_prisma.prisma.hook_point import HookPoint
+from vit_prisma.prisma_tools.hook_point import HookPoint
 
 import logging
